Chore: reorganize config modules
