+++ src/careamics/config/__init__.py
@@ -41,7 +41,6 @@ from .algorithms import (
     UNetBasedAlgorithm,
     VAEBasedAlgorithm,
 )
-from .callback_model import CheckpointModel
 from .configuration import Configuration
 from .configuration_factories import (
     algorithm_factory,
@@ -51,9 +50,13 @@ from .configuration_factories import (
     create_n2n_configuration,
     create_n2v_configuration,
 )
-from .configuration_io import load_configuration, save_configuration
 from .data import DataConfig
-from .inference_model import InferenceConfig
-from .loss_model import LVAELossConfig
-from .nm_model import GaussianMixtureNMConfig, MultiChannelNMConfig
-from .training_model import TrainingConfig
+from .data.inference_model import InferenceConfig
+from .lightning.callbacks.callback_model import CheckpointModel
+from .lightning.training_model import TrainingConfig
+from .losses.loss_model import LVAELossConfig
+from .noise_model.noise_model_config import (
+    GaussianMixtureNMConfig,
+    MultiChannelNMConfig,
+)
+from .utils.configuration_io import load_configuration, save_configuration
