+++ src/careamics/config/configuration.py
@@ -18,7 +18,7 @@ from careamics.config.algorithms import (
     N2VAlgorithm,
 )
 from careamics.config.data import DataConfig
-from careamics.config.training_model import TrainingConfig
+from careamics.config.lightning.training_model import TrainingConfig
 from careamics.lvae_training.dataset.config import MicroSplitDataConfig
 
 ALGORITHMS = Union[
