+++ src/careamics/config/data/tile_information.py
