+++ src/careamics/config/lightning/__init__.py
@@ -1,0 +1,1 @@
+"""Training and lightning related Pydantic configurations."""
