+++ src/careamics/config/lightning/callbacks/__init__.py
@@ -1,0 +1,1 @@
+"""Callbacks Pydantic configurations."""
