+++ src/careamics/config/lightning/callbacks/callback_model.py
