+++ src/careamics/config/losses/__init__.py
@@ -1,0 +1,1 @@
+"""Losses Pydantic configurations."""
