+++ src/careamics/config/noise_model/__init__.py
@@ -1,0 +1,1 @@
+"""Noise models Pydantic configurations."""
