+++ src/careamics/config/noise_model/likelihood_model.py
