+++ src/careamics/config/noise_model/noise_model_config.py
