+++ src/careamics/config/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Configuration utilities."""
