+++ src/careamics/dataset/iterable_tiled_pred_dataset.py
@@ -13,7 +13,7 @@ from careamics.file_io.read import read_tiff
 from careamics.transforms import Compose
 
 from ..config import InferenceConfig
-from ..config.tile_information import TileInformation
+from ..config.data.tile_information import TileInformation
 from ..config.transformations import NormalizeModel
 from .dataset_utils import iterate_over_files
 from .tiling import extract_tiles
