+++ src/careamics/dataset_ng/legacy_interoperability.py
@@ -9,7 +9,7 @@ from typing import cast
 import numpy as np
 from numpy.typing import NDArray
 
-from careamics.config.tile_information import TileInformation
+from careamics.config.data.tile_information import TileInformation
 
 from .dataset import ImageRegionData
 from .patching_strategies import TileSpecs
