+++ src/careamics/lightning/lightning_module.py
@@ -13,6 +13,7 @@ from careamics.config import (
     VAEBasedAlgorithm,
     algorithm_factory,
 )
+from careamics.config.data.tile_information import TileInformation
 from careamics.config.support import (
     SupportedAlgorithm,
     SupportedArchitecture,
@@ -20,7 +21,6 @@ from careamics.config.support import (
     SupportedOptimizer,
     SupportedScheduler,
 )
-from careamics.config.tile_information import TileInformation
 from careamics.losses import loss_factory
 from careamics.models.lvae.likelihoods import (
     GaussianLikelihood,
