+++ src/careamics/prediction_utils/stitch_prediction.py
@@ -6,7 +6,7 @@ from typing import Union
 import numpy as np
 from numpy.typing import NDArray
 
-from careamics.config.tile_information import TileInformation
+from careamics.config.data.tile_information import TileInformation
 
 
 class TilingMode:
