+++ tests/config/test_inference_model.py
@@ -1,6 +1,6 @@
 import pytest
 
-from careamics.config.inference_model import InferenceConfig
+from careamics.config.data.inference_model import InferenceConfig
 
 
 @pytest.mark.parametrize("ext", ["nd2", "jpg", "png ", "zarr", "npy"])
