+++ tests/config/test_tile_information.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from careamics.config.tile_information import TileInformation
+from careamics.config.data.tile_information import TileInformation
 
 
 def test_defaults():
