+++ tests/dataset/tiling/test_lvae_tiled_patching.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from careamics.config.tile_information import TileInformation
+from careamics.config.data.tile_information import TileInformation
 from careamics.dataset.tiling.lvae_tiled_patching import (
     compute_padding,
     compute_tile_grid_shape,
